requirements.txt
opencv-python==4.9.0.80
matplotlib==3.8.4
numpy==1.26.4
pillow==10.3.0
pandas==2.2.2
seaborn==0.13.2
scikit-learn==1.4.2
albumentations==1.4.8
albucore==0.0.16
torch==2.3.0
notebook==7.2.0
streamlit==1.28.1
requests==2.31.0
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
ultralytics==8.0.196
Pillow==10.0.1
easyocr==1.7.0
python-jose[cryptography]==3.3.0
mlflow==2.12.0